+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The following changes have been completed, and will be included in the next tagg
 * Added `AnyServiceManager` type alias
 * Added `aliasFor()` ServiceProvider builder
 * Added `existingService()` ServiceProvider builder
+* Added `sharedInstance()` ServiceProvider builder
 
 ### Fixes